chore: dispose when session is terminated
